# docs/swagger.yaml
definitions:
  common.Pagination:
    properties:
      currentPage:
        description: 當前頁碼
        example: 1
        type: integer
      keywords:
        description: 關鍵字
        example: software
        type: string
      pageSize:
        description: 每頁大小
        example: 10
        type: integer
      sortBy:
        example: name
        type: string
      totalPages:
        description: 總頁數
        type: integer
      totalRecords:
        description: 總記錄數
        type: integer
    type: object
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  model.Individual:
    properties:
      created_at:
        type: string
      email:
        example: john@example.com
        type: string
      head_img_path:
        example: /images/john_doe.jpg
        type: string
      id:
        type: string
      job_title:
        example: Software Engineer
        type: string
      name:
        example: John Doe
        type: string
      projects:
        items:
          $ref: '#/definitions/model.Project'
        type: array
      resume_link:
        example: https://example.com/john_doe_resume.pdf
        type: string
      skills:
        example:
        - Golang
        - JavaScript
        - Python
        items:
          type: string
        type: array
      social_media:
        $ref: '#/definitions/model.SocialMedia'
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.Project:
    properties:
      demo_link:
        example: https://example.com/project_x_demo
        type: string
      github_rep_link:
        example: https://github.com/johndoe/project_x
        type: string
      id:
        type: string
      introduce:
        example: A project for implementing microservices architecture.
        type: string
      name:
        example: Project X
        type: string
      skills:
        example:
        - '[''Golang'''
        - ' ''Docker'''
        - ' ''Kubernetes'']'
        items:
          type: string
        type: array
      user_id:
        type: string
    type: object
  model.SocialMedia:
    properties:
      facebook:
        example: johndoe
        type: string
      github:
        example: johndoe
        type: string
      linkedin:
        example: johndoe
        type: string
      user_id:
        type: string
    type: object
  model.User:
    properties:
      createdAt:
        description: 創建時間
        type: string
      disable:
        description: 帳號是否註銷
        type: integer
      email:
        description: 電子信箱
        type: string
      headerImg:
        description: 會員頭像連結
        type: string
      id:
        description: 主鍵ID
        type: integer
      last_login_at:
        type: string
      nickName:
        type: string
      phone:
        description: 手機號碼
        type: string
      updatedAt:
        description: 更新時間
        type: string
      userName:
        description: 會員名稱
        type: string
      uuid:
        description: 會員UUID
        type: string
    type: object
  request.Login:
    properties:
      account:
        description: 帳號
        example: "0912345678"
        type: string
      password:
        description: 密碼
        example: "123456"
        type: string
    required:
    - account
    - password
    type: object
  request.Register:
    properties:
      email:
        example: 123@gmail.com
        type: string
      headerImg:
        example: http://headimgurl.com
        type: string
      nickName:
        example: Ed
        type: string
      passWord:
        example: "123456"
        type: string
      phone:
        example: "0912345678"
        type: string
      userName:
        example: Edward
        type: string
    required:
    - email
    - passWord
    - phone
    - userName
    type: object
  response.LoginResponse:
    properties:
      expiresAt:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
  response.UserResponse:
    properties:
      user:
        $ref: '#/definitions/model.User'
    type: object
info:
  contact: {}
  description: 使用golang gin開發全端個人作品集網站
  title: E-Portfolio Swagger API接口文件
  version: v1.0.0
paths:
  /v1/auth/delete:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 註銷帳號
      tags:
      - Auth
  /v1/auth/login:
    post:
      parameters:
      - description: 信箱或手機號碼, 密碼
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: 返回使用者資訊,token,過期時間
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.LoginResponse'
                msg:
                  type: string
              type: object
      summary: 會員登入
      tags:
      - Auth
  /v1/auth/refresh-token:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 刷新token
      tags:
      - Auth
  /v1/auth/register:
    post:
      parameters:
      - description: 使用者名稱, 密碼, 暱稱, 頭貼網址, 手機號碼, 信箱
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: 會員註冊,返回會員註冊訊息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
                msg:
                  type: string
              type: object
      summary: 會員註冊
      tags:
      - Auth
  /v1/individual/create:
    post:
      parameters:
      - description: 個人作品集資料
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Individual'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 新增個人作品集資料
      tags:
      - Individual
  /v1/individual/update:
    post:
      parameters:
      - description: 更新個人作品集資料
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Individual'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 更新個人作品集資料
      tags:
      - Individual
  /v1/portfolio/{userName}:
    get:
      parameters:
      - description: user Name
        in: path
        name: userName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 個人作品集
      tags:
      - Portfolio
  /v1/portfolio/list:
    post:
      parameters:
      - description: 取得分頁列表
        in: body
        name: data
        schema:
          $ref: '#/definitions/common.Pagination'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 個人作品集列表
      tags:
      - Portfolio
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"